Module nodle_chain_runtime::constants

Structs

Perbill
A fixed point representation of a number in the range [0, 1].

Perquintill
A fixed point representation of a number in the range [0, 1].

RuntimeBlockWeights

Constants

AVERAGE_ON_INITIALIZE_RATIO
We assume that ~10% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

CENTS

DAYS

DOLLARS

EPOCH_DURATION_IN_BLOCKS

EPOCH_DURATION_IN_SLOTS

HOURS

MAXIMUM_BLOCK_WEIGHT
We allow for 2 seconds of compute with a 6 second average block time.

MILLICENTS

MILLISECS_PER_BLOCK
Time and blocks.

MINUTES

NODL
Money matters.

NORMAL_DISPATCH_RATIO
We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

PRIMARY_PROBABILITY

SLOT_DURATION

TARGET_BLOCK_FULLNESS
Fee-related. The block saturation level. Fees will be updates based on this value.

Functions

deposit